Move CustomerNameContainsKeywordsPredicate into package
