+++ src/main/java/seedu/address/model/customer/predicates/CustomerNameContainsKeywordsPredicate.java
@@ -1,9 +1,10 @@
-package seedu.address.model.customer;
+package seedu.address.model.customer.predicates;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
+import seedu.address.model.customer.Customer;
 
 /**
  * Tests that a {@code Customer}'s {@code Name} matches any of the keywords given.
