+++ src/test/java/seedu/address/model/customer/CustomerNameContainsKeywordsPredicateTest.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.customer.predicates.CustomerNameContainsKeywordsPredicate;
 import seedu.address.testutil.CustomerBuilder;
 
 public class CustomerNameContainsKeywordsPredicateTest {
